address comments, fix batch_info false population to () from None in algorithm.py.
