+++ alf/algorithms/data_transformer_test.py
@@ -13,13 +13,16 @@
 # limitations under the License.
 
 from absl.testing import parameterized
+import gin
 import torch
 
 import alf
 from alf.data_structures import namedtuple, StepType
 from alf.experience_replayers.replay_buffer import ReplayBuffer, BatchInfo
-from alf.algorithms.data_transformer import FrameStacker, ImageScaleTransformer
+from alf.experience_replayers.replay_buffer_test import ReplayBufferTest
+from alf.algorithms.data_transformer import FrameStacker, ImageScaleTransformer, HindsightExperienceTransformer
 from alf.utils import common
+from alf.utils.data_buffer_test import get_batch
 
 DataItem = namedtuple(
     'DataItem', ['step_type', 'observation', 'batch_info', 'replay_buffer'],
@@ -191,6 +194,90 @@ class ImageScaleTransformerTest(alf.test.TestCase):
             min=0.,
             fields=['other'])
         transformer = ImageScaleTransformer(spec, min=0., fields=['img'])
+
+
+class HindsightExperienceTransformerTest(ReplayBufferTest):
+    @parameterized.named_parameters([
+        ('test_dense_epi_ends', 0.1),
+        ('test_sparse_epi_ends', 0.004),
+    ])
+    def test_compute_her_future_step_distance(self, end_prob):
+        num_envs = 2
+        max_length = 100
+        torch.manual_seed(0)
+        configs = [
+            "HindsightExperienceTransformer.her_proportion=0.8",
+            'HindsightExperienceTransformer.achieved_goal_field="o.a"',
+            'HindsightExperienceTransformer.desired_goal_field="o.g"'
+        ]
+        gin.parse_config_files_and_bindings("", configs)
+
+        replay_buffer = ReplayBuffer(
+            data_spec=self.data_spec,
+            num_environments=num_envs,
+            max_length=max_length,
+            keep_episodic_info=True,
+            step_type_field="t")
+
+        transform = HindsightExperienceTransformer(self.data_spec)
+
+        # insert data
+        max_steps = 1000
+        # generate step_types with certain density of episode ends
+        steps = self.generate_step_types(
+            num_envs, max_steps, end_prob=end_prob)
+        for t in range(max_steps):
+            for b in range(num_envs):
+                batch = get_batch([b],
+                                  self.dim,
+                                  t=steps[b * max_steps + t],
+                                  x=1. / max_steps * t + b)
+                replay_buffer.add_batch(batch, batch.env_id)
+            if t > 1:
+                sample_steps = min(t, max_length)
+                env_ids = torch.tensor([0] * sample_steps + [1] * sample_steps)
+                idx = torch.tensor(
+                    list(range(sample_steps)) + list(range(sample_steps)))
+                gd = self.steps_to_episode_end(replay_buffer, env_ids, idx)
+                idx_orig = replay_buffer._indexed_pos.clone()
+                idx_headless_orig = replay_buffer._headless_indexed_pos.clone()
+                d = replay_buffer.steps_to_episode_end(
+                    replay_buffer._pad(idx, env_ids), env_ids)
+                # Test distance to end computation
+                if not torch.equal(gd, d):
+                    outs = [
+                        "t: ", t, "\nenvids:\n", env_ids, "\nidx:\n", idx,
+                        "\npos:\n",
+                        replay_buffer._pad(idx, env_ids), "\nNot Equal: a:\n",
+                        gd, "\nb:\n", d, "\nsteps:\n", replay_buffer._buffer.t,
+                        "\nindexed_pos:\n", replay_buffer._indexed_pos,
+                        "\nheadless_indexed_pos:\n",
+                        replay_buffer._headless_indexed_pos
+                    ]
+                    outs = [str(out) for out in outs]
+                    assert False, "".join(outs)
+
+                # Save original exp for later testing.
+                g_orig = replay_buffer._buffer.o["g"].clone()
+                r_orig = replay_buffer._buffer.reward.clone()
+
+                # HER relabel experience
+                res, info = replay_buffer.get_batch(sample_steps, 2)
+                res = res._replace(batch_info=info)
+                res = transform.transform_experience(res)
+
+                self.assertEqual(list(res.o["g"].shape), [sample_steps, 2])
+
+                # Test relabeling doesn't change original experience
+                self.assertTrue(
+                    torch.allclose(r_orig, replay_buffer._buffer.reward))
+                self.assertTrue(
+                    torch.allclose(g_orig, replay_buffer._buffer.o["g"]))
+                self.assertTrue(
+                    torch.all(idx_orig == replay_buffer._indexed_pos))
+                self.assertTrue(
+                    torch.all(idx_headless_orig == replay_buffer.
+                              _headless_indexed_pos))
 
 
 if __name__ == '__main__':
